Add warning for zero portions or weight and set default weight
This commit introduces a warning message that appears if the user tries to calculate ingredients with zero portions or zero portion weight.
The warning message "Number of portions or portion weight can't be zero" is added in English, Spanish, Italian, and Catalan.

Additionally, the default value for portion weight is set to 280 in `MainActivity.java`.