Add validation for empty file name in `SaveFileDialog`
This commit introduces a check in `SaveFileDialog.java` to prevent saving a file with an empty name.
If the user tries to save with an empty file name, a Toast message "File name can't be empty" (localized) is displayed.

The following string resources were added:
- `toast_file_name_empty` in `values/strings.xml` (English and Spanish translations provided, Catalan and Italian also updated)

Additionally, a minor typo correction was made to `toast_portions_warning` in `values-ca-rES/strings.xml`.